Migrate scripts into subdirectory
